update view model
