+++ Tajanb/TajanbApp.swift
@@ -10,36 +10,13 @@ import SwiftData
 
 @main
 struct TajanbApp: App {
-    // Initialize the view models
-    let textRecognitionViewModel = TextRecognitionViewModel(categoryManager: CategoryManagerViewModel())
-    let categoryManagerViewModel = CategoryManagerViewModel()
-
+    let viewModel = CameraViewModel()
+    let photoviewModel = PhotoViewModel(viewmodel: CameraViewModel())
     var body: some Scene {
         WindowGroup {
             CameraView(
-                textRecognitionViewModel: textRecognitionViewModel,
-                categoryManagerViewModel: categoryManagerViewModel, photoViewModel: PhotoViewModel(categoryManager: categoryManagerViewModel)
-            )
+                viewModel: viewModel, photoViewModel: photoviewModel)
         }
     }
 }
-//    var sharedModelContainer: ModelContainer = {
-//        let schema = Schema([
-//            Item.self,
-//        ])
-//        let modelConfiguration = ModelConfiguration(schema: schema, isStoredInMemoryOnly: false)
-//
-//        do {
-//            return try ModelContainer(for: schema, configurations: [modelConfiguration])
-//        } catch {
-//            fatalError("Could not create ModelContainer: \(error)")
-//        }
-//    }()
-//
-//    var body: some Scene {
-//        WindowGroup {
-//            ContentView()
-//        }
-//        .modelContainer(sharedModelContainer)
-//    }
-//}
+
